day-02-tiny-image-captioner/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
httpx

